#60;support strong type field

+++ java/src/main/java/com/taobao/top/link/endpoint/Message.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 // just simple version
 public class Message {
-	public short protocolVersion = 1;
+	public int protocolVersion = 1;
 	public short messageType;
 
 	public int statusCode;
